Fix race condition in refresh by not connecting until resolver is complete.
The resolver needs to run before the place to connect is determined, but
node-redis will start attempting to reconnect immediately. Prevent this by
setting the port and ip to the empty string and letting the resolver
modify it.

A better fix would be to modify node-redis to be able to not retry until
told to do so.